requirements.txt
flask
langchain
langchain-core
langchain-google-genai
langgraph
pydantic
python-dotenv
werkzeug

